Skuli

+++ js/view.js
@@ -2,9 +2,10 @@ function View() {
     "use strict";
     
     var table = document.getElementById('fruitsTable');
+    
     var updateFullTable = function(){
          var fruits = controller.show();
-        
+
         $("#fruitsTable tr").remove();
         
         for(var i=0; i < fruits.length; i++){
@@ -44,20 +45,26 @@ function View() {
         updateFullTable();
     }
     
+    buttonSearch.onclick = function() {
+        var item = document.getElementById('search').value;
+        controller.search(item);
+        updateFullTable();
+    }
+    
+    buttonClearSearch.onclick = function() {
+        controller.clearSearch();
+        updateFullTable();
+    }
+    
     buttonAdd.onclick=function(){
         
         addNodeToTable();
     }
         
     buttonDelete.onclick=function(){
-        var textInput = document.getElementById('textInput').value;
-        var oldLength = controller.show().length;
-
-        controller.deleteByName(textInput);
-
-        var newLength = controller.show().length;
-        alert('old = '+ oldLength + ' new = ' + newLength);
-
+        var textInput = document.getElementById('textDelete').value;
+        if(textInput != '')
+            controller.deleteByName(textInput);
         updateFullTable(); 
     }
         
@@ -72,4 +79,37 @@ function View() {
 
         updateFullTable(); 
     }
-}
+    
+    buttonEditSearch.onclick = function() {
+        var textInput = document.getElementById('editSearchInput').value;
+        console.log(textInput);
+        var inputs = [];
+        inputs = controller.searchToEdit(textInput);
+        
+        if(inputs[0] != undefined || inputs.length !== 0 && textInput != undefined){
+            document.getElementById('editNameInput').value = inputs[0];
+            document.getElementById('editPriceInput').value = inputs[1];
+            document.getElementById('editOriginInput').value = inputs[2];
+            }
+    }
+    
+    buttonEdit.onclick = function(){
+        var fruitTodelete = document.getElementById('editSearchInput').value;
+        var fruitName = document.getElementById('editNameInput').value;
+        var price = document.getElementById('editPriceInput').value;            var origin = document.getElementById('editOriginInput').value;
+        
+        if(fruitName != "" && price != "" && origin != "") {
+            if(fruitName != undefined && price != undefined && origin != undefined)
+                   controller.deleteByName(fruitTodelete);
+        controller.add(fruitName, price, origin); 
+        }
+
+        updateFullTable();
+    }
+    
+    this.validateInput = function(){
+        
+    }
+      
+  }
+
